feat: Implement Admins Management Table and Form
- Added AdminsTable component for displaying and managing admin users.
- Integrated search functionality for filtering users by name or email.
- Implemented add, edit, block, and unblock functionalities for admin users.
- Created CategoryForm component for adding and editing categories with image upload.
- Developed CategoryTable component for displaying categories with options to edit and delete.
- Introduced Sliders component for managing slider images with upload and delete options.
- Enhanced Dashboard, HomePage, and other pages with respective components for better management.
- Updated routing and context usage for improved state management across components.

+++ utils/requestsUtils.js
@@ -1,24 +1,47 @@
 import axios from "axios";
+import Swal from "sweetalert2";
 
-export const postRequest = async (endpoint, data) => {
+export const postRequest = async (endpoint, dataBody, message) => {
   const token =
     typeof window !== "undefined" ? localStorage.getItem("accessToken") : null;
   // const lang =
   //   typeof window !== "undefined" ? localStorage.getItem("lang") : null;
   try {
-    const response = await axios.post(
-      process.env.NEXT_PUBLIC_API_BASE_URL + endpoint,
-      data,
-      {
-        headers: {
-        //  " Content-Type": "multipart/form-data",
-          Authorization: token ? `Bearer ${token}` : undefined,
-          "Accept-Language": localStorage.getItem("lang"),
-        },
-      }
-    );
-    console.log(response);
-    return await response.data;
+    
+    if (message === "") {
+      const response = await axios.post(
+        process.env.NEXT_PUBLIC_API_BASE_URL + endpoint,
+        dataBody,
+        {
+          headers: {
+            Authorization: token ? `Bearer ${token}` : undefined,
+            "Accept-Language": localStorage.getItem("lang"),
+          },
+        }
+      );
+      console.log(response);
+      return await response.data;
+    } else {
+     await Swal.fire({
+        text: message,
+        showCancelButton: true,
+      }).then(async (data) => {
+        if (data.isConfirmed) {
+          const response = await axios.post(
+            process.env.NEXT_PUBLIC_API_BASE_URL + endpoint,
+            dataBody,
+            {
+              headers: {
+                Authorization: token ? `Bearer ${token}` : undefined,
+                "Accept-Language": localStorage.getItem("lang"),
+              },
+            }
+          );
+          console.log(response);
+          return await response.data;
+        }
+      });
+    }
   } catch (error) {
     console.error("POST request error:", error);
     throw error;
@@ -28,7 +51,6 @@ export const postRequest = async (endpoint, data) => {
 export const getRequest = async (endpoint) => {
   const token =
     typeof window !== "undefined" ? localStorage.getItem("accessToken") : null;
-
 
   try {
     const response = await axios.get(
@@ -50,25 +72,31 @@ export const getRequest = async (endpoint) => {
   }
 };
 
-export const putRequest = async (endpoint, data) => {
+export const putRequest = async (endpoint, dataBody, message) => {
   const token =
     typeof window !== "undefined" ? localStorage.getItem("accessToken") : null;
 
-
   try {
-    const response = await axios.put(
-      process.env.NEXT_PUBLIC_API_BASE_URL + endpoint,
-      data,
-      {
-        headers: {
-          // "Content-Type": "multipart/form-data",
-          Authorization: token ? `Bearer ${token}` : undefined,
-          "Accept-Language": localStorage.getItem("lang"),
-        },
+  await  Swal.fire({
+      text: message,
+      showCancelButton: true,
+    }).then(async (data) => {
+      if (data.isConfirmed) {
+        const response = await axios.put(
+          process.env.NEXT_PUBLIC_API_BASE_URL + endpoint,
+          dataBody,
+          {
+            headers: {
+              // "Content-Type": "multipart/form-data",
+              Authorization: token ? `Bearer ${token}` : undefined,
+              "Accept-Language": localStorage.getItem("lang"),
+            },
+          }
+        );
+        console.log(response);
+        return await response.data;
       }
-    );
-    console.log(response);
-    return await response.data;
+    });
   } catch (error) {
     console.error("POST request error:", error);
     throw error;
@@ -78,7 +106,6 @@ export const putRequest = async (endpoint, data) => {
 // export const patchRequest = async (endpoint, data) => {
 //   const token =
 //     typeof window !== "undefined" ? localStorage.getItem("accessToken") : null;
-
 
 //   try {
 //     const response = await axios.patch(
@@ -99,24 +126,39 @@ export const putRequest = async (endpoint, data) => {
 //   }
 // };
 
-export const deleteRequest = async (endpoint, data) => {
+export const deleteRequest = async (endpoint, message) => {
   const token =
     typeof window !== "undefined" ? localStorage.getItem("accessToken") : null;
 
-
   try {
-    const response = await axios.delete(
-      process.env.NEXT_PUBLIC_API_BASE_URL + endpoint,
-      {
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: token ? `Bearer ${token}` : undefined,
-          "Accept-Language": "ar",
-        },
+   await Swal.error({
+    title:localStorage.lang ==='en'? message +'Delete':message+"حذف",
+      showCancelButton: true,
+      confirmButtonText: localStorage.lang === "ar" ? "موافق" : "OK",
+      cancelButtonText: localStorage.lang === "ar" ? "إلغاء" : "Cancel",
+      customClass: {
+      
+    confirmButton: 'bg-blue-600 hover:  bg-blue-500 text-white px-4 py-2 rounded-lg font-semibold',
+    cancelButton: 'bg-gray-300 hover:bg-gray-400 text-gray-800 px-4 py-2 rounded-lg'        
+      },
+     
+      
+    }).then(async (data) => {
+      if (data.isConfirmed) {
+        const response = await axios.delete(
+          process.env.NEXT_PUBLIC_API_BASE_URL + endpoint,
+          {
+            headers: {
+              "Content-Type": "application/json",
+              Authorization: token ? `Bearer ${token}` : undefined,
+              "Accept-Language": "ar",
+            },
+          }
+        );
+        console.log(response);
+        return await response.data;
       }
-    );
-    console.log(response);
-    return await response.data;
+    });
   } catch (error) {
     console.error("POST request error:", error);
     throw error;
